requirements.txt
mysql-connector-python
elasticsearch

requests